Tweak auth setup
