+++ api/src/auth.ts
@@ -1,0 +1,37 @@
+import env from './env';
+import { toArray } from './utils/to-array';
+import { Issuer } from 'openid-client';
+
+// oAuth 2 /////////////////////////////////////////////////////////////////////////////////////////
+const oauthProviders = toArray(env.OAUTH_PROVIDERS).map((provider) => provider.toLowerCase());
+
+export const oauthConfig: any = {
+	defaults: {
+		origin: env.PUBLIC_URL,
+		transport: 'session',
+		prefix: '/auth/oauth',
+		response: ['tokens', 'profile'],
+	},
+};
+
+for (const [key, value] of Object.entries(env)) {
+	if (key.startsWith('OAUTH') === false) continue;
+
+	const parts = key.split('_');
+	const provider = parts[1].toLowerCase();
+
+	if (oauthProviders.includes(provider) === false) continue;
+
+	// OAUTH <PROVIDER> SETTING = VALUE
+	parts.splice(0, 2);
+
+	const configKey = parts.join('_').toLowerCase();
+
+	oauthConfig[provider] = {
+		...(oauthConfig[provider] || {}),
+		[configKey]: value,
+	};
+}
+
+// Open-ID Connect (OICD) //////////////////////////////////////////////////////////////////////////
+export async function registerOIDC() {}
